requirements.txt
python-telegram-bot[webhooks]
openai
langchain
langgraph
langchain-openai
langchain-teddynote
psycopg2-binary